requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6

# Database
motor==3.3.2
beanie==1.23.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.2.1
httpx==0.25.2

# LLM Providers
openai==1.3.0
aiohttp==3.9.1

# Templates
jinja2==3.1.2
jsonschema==4.19.2

# Background tasks
celery==5.3.4
kombu==5.3.4

# Export formats
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
lxml==4.9.3
markdown==3.5.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0
mongomock-motor==0.0.21

# Development
python-dotenv==1.0.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1
pre-commit==3.5.0